Fix the load/save/erase game menu.
First, Asyncify seems to have some kind of limitation with respect to the length or complexity of the functions you're using, or something.
We fix a crash by breaking the function up a little bit.

Secondly, we make readcontrols() sleep for a moment so that input events can roll in.  This fixes some cases where the original menu code hangs.